+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 * refactor and export response builder.
+* support objects and arrays as `bodyPatch` values.
 
 ## [0.3.0] - 2026-05-01
 * remove the `defaultMethod` option from `MockClient`.